Properly pass config object.
We do pass the config object so we can use it for things like proxy or
SSL verification when doing requests. In a call to
`doArtifactoryRequest` it was not properly passed so the settings were
not used in one of the error case handlers.

This patch fixes the call to `doArtifactoryRequest`.